test: split MP4 integration tests by concern
